Structuur beetje aangepast, en vooral geprobeert interacties helemaal toevoegen (wat gelukt is) alleen wordt er nog gekeken naar animator runtime

+++ Interaction-layer/Assets/Software/Business layer/Domain/Hardware.cs
@@ -11,9 +11,9 @@ namespace Business.Domain {
 		public List<Interaction> interactions;
 		public State state;
 		public Type type;
-		public int x; 
-		public int y; 
-		public int z;
+		public float x; 
+		public float y; 
+		public float z;
 		public Hardware() {
 		}
 		public override string ToString ()
